changed hours_played in database creation file to float to account for decimal precision

+++ sql-create-tables.sql
@@ -14,7 +14,7 @@ CREATE EXTERNAL TABLE dbo.games_ranking (
 
 CREATE EXTERNAL TABLE dbo.steam_game_reviews (
     id INT IDENTITY(1,1),
-    hours_played int,
+    hours_played float,
     recommendation VARCHAR(255),
     `date` DATE,
     game_name VARCHAR(255),
